Corrected pixel overflow
The unsigned int, does cover more of the data as expected.
In the original code, this was treated by a two byte storage,
as the new storage is four byte long, to much data is written.
It would be possible, to use the same mechanism here, which
would even save a bit of memory. But I think using the mask
is faster as the handling of the memory offset. Not measured
it so the two byte solution might be still faster

Anyhow both solutions are working